done for ex01

+++ ex00/megaphone.cpp
@@ -9,8 +9,8 @@ int main(int ac, char *argv[])
     for (int i = 1; i < ac; i++)
     {
         str = argv[i];
-        for(int j = 0; j < str.length(); j++)
-            std::cout<<(char)std::toupper(str[j]);
+        for(size_t j = 0; j < str.length(); j++)
+            std::cout<< (char)std::toupper(str[j]);
     }
     std::cout<<std::endl;
     return (0);
